# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api of kindergarden",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parents": {
            "get": {
                "tags": [
                    "Parents"
                ],
                "summary": "All parents",
                "description": "All parents",
                "operationId": "all_parents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "identification": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "parent": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "residence_phone": {
                                            "type": "string",
                                            "format": "string",
                                            "example": ""
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "military": {
                                            "type": "boolean",
                                            "format": "boolean",
                                            "example": ""
                                        },
                                        "institution": {
                                            "type": "string",
                                            "format": "string",
                                            "example": ""
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Parents"
                ],
                "summary": "Store parent",
                "description": "Store parent",
                "operationId": "store_parents",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Peralta"
                                    },
                                    "identification": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "40224522776"
                                    },
                                    "parent": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Primo"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "8099877765"
                                    },
                                    "residence_phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "8099886700"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Santo Domingo"
                                    },
                                    "military": {
                                        "type": "boolean",
                                        "format": "boolean",
                                        "example": "true"
                                    },
                                    "institution": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "ARD"
                                    },
                                    "kid_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/parents/{id}": {
            "get": {
                "tags": [
                    "Parents"
                ],
                "summary": "See parent",
                "description": "See parent",
                "operationId": "watch_parents",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "identification": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "parent": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "residence_phone": {
                                            "type": "string",
                                            "format": "string",
                                            "example": ""
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "military": {
                                            "type": "boolean",
                                            "format": "boolean",
                                            "example": "true"
                                        },
                                        "institution": {
                                            "type": "string",
                                            "format": "string",
                                            "example": ""
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Parents"
                ],
                "summary": "Update parent",
                "description": "Update parent",
                "operationId": "update_parents",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Peralta"
                                    },
                                    "identification": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "40224522776"
                                    },
                                    "parent": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Primo"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "8099877765"
                                    },
                                    "residence_phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "8099886700"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Santo Domingo"
                                    },
                                    "military": {
                                        "type": "boolean",
                                        "format": "boolean",
                                        "example": "true"
                                    },
                                    "institution": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "ARD"
                                    },
                                    "kid_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Parents"
                ],
                "summary": "Delete parent",
                "description": "Delete parent",
                "operationId": "delete_parents",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    }
}